third_party/WebKit/LayoutTests/inspector/sources/debugger-frameworks/frameworks-sourcemap.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script src="../debugger/resources/framework-with-sourcemap.js"></script>
<script>

function testFunction()
{
    debugger;
    return foo(callback);
}

function callback(i)
{
    return i;
}

function test()
{
    TestRunner.addSniffer(Bindings.BlackboxManager.prototype, "_patternChangeFinishedForTests", step1);
    var frameworkRegexString = "/framework\\.js$";
    Common.settingForTest("skipStackFramesPattern").set(frameworkRegexString);

    function step1()
    {
        SourcesTestRunner.startDebuggerTest(step2, true);
    }

    function step2()
    {
        SourcesTestRunner.runTestFunctionAndWaitUntilPaused(step3);
    }

    function step3()
    {
        var actions = [
            "Print", // "debugger" in testFunction()
            "StepInto", "StepInto", "Print", // entered callback(i)
            "StepOut", "Print"
        ];
        SourcesTestRunner.waitUntilPausedAndPerformSteppingActions(actions, step4);
    }

    function step4()
    {
        SourcesTestRunner.completeDebuggerTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests framework blackboxing feature with sourcemaps.</p>
</body>
</html>
