third_party/WebKit/LayoutTests/inspector/sources/debugger-frameworks/frameworks-sourcemap-expected.txt

Tests framework blackboxing feature with sourcemaps.

Set timer for test function.
Call stack:
    0) testFunction (frameworks-sourcemap.html:10)

Executing StepInto...
Executing StepInto...
Call stack:
    0) callback (frameworks-sourcemap.html:16)
  * 1) foo (framework.js:4)
    2) testFunction (frameworks-sourcemap.html:11)

Executing StepOut...
Call stack:
    0) testFunction (frameworks-sourcemap.html:12)
